% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_utils.R
\name{adjustForDiscrete}
\alias{adjustForDiscrete}
\title{Adjust the correlation matrix when there are discrete distributions present}
\usage{
adjustForDiscrete(rho, params, nSigmas)
}
\arguments{
\item{rho}{The input correlation matrix}

\item{params}{The parameters of the marginals.}

\item{nSigmas}{The number of standard deviations from the mean}
}
\description{
Adjust the correlation matrix when there are discrete distributions present
}
