% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{adjustInputR}
\alias{adjustInputR}
\title{Adjust an input correlation matrix for our method to match}
\usage{
adjustInputR(R, params, cores = 1)
}
\arguments{
\item{R}{A a square symmetric correlation matrix.}

\item{params}{The parameters of the marginals.}

\item{cores}{The number of cores to utilize.}
}
\description{
Adjust an input correlation matrix for our method to match
}
