% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_utils.R
\name{all_corInBounds}
\alias{all_corInBounds}
\title{Returns TRUE if all correlations pairs are within the theoretical bounds}
\usage{
all_corInBounds(
  rho,
  margins,
  cores = 1,
  type = c("pearson", "spearman", "kendall"),
  rho_bounds = NULL,
  ...
)
}
\arguments{
\item{rho}{The input correlation matrix.}

\item{margins}{The parameters of the marginals.}

\item{cores}{The number of cores to utilize.}

\item{type}{The type of correlation matrix that is being passed.}

\item{rho_bounds}{Pre-computed upper and lower correlation bounds
... Other arguments passed to `computeCoreBounds()`}
}
\value{
Logical. TRUE if all correlations pairs are within the theoretical
  bounds, and false otherwise.
}
\description{
Returns TRUE if all correlations pairs are within the theoretical bounds
}
