% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{brca}
\alias{brca}
\title{BRCA dataset}
\format{
An object of class \code{data.frame} with 878 rows and 200 columns.
}
\usage{
brca
}
\description{
200 of the highest expressing genes from the BRCA data set.
}
\keyword{datasets}
