% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_utils.R
\name{computeCorBounds}
\alias{computeCorBounds}
\title{Computes the theoretical upper and lower bounds of possible correlations
  given a set of marginals}
\usage{
computeCorBounds(params, cores = 1, type = c("pearson", "kendall",
  "spearman"), reps = 1e+05)
}
\arguments{
\item{params}{The parameters of the marginals.}

\item{cores}{The number of cores to utilize.}

\item{type}{The type of correlation matrix that is being passed.}
}
\value{
A list containing the theoretical upper and lower bounds
}
\description{
Computes the theoretical upper and lower bounds of possible correlations
  given a set of marginals
}
