% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_utils.R
\name{constrainRho}
\alias{constrainRho}
\title{Constrains a correlation matrix to the theoretical upper and lower bounds}
\usage{
constrainRho(rho, rho_bounds)
}
\arguments{
\item{rho}{The input correlation matrix.}

\item{rho_bounds}{A list containing the theoretical upper and lower bounds}
}
\value{
The constrained correlation matrix
}
\description{
Constrains a correlation matrix to the theoretical upper and lower bounds
}
