% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_utils.R
\name{convertCor}
\alias{convertCor}
\title{Convert between different types of correlation matrices}
\usage{
convertCor(rho, from = c("pearson", "spearman", "kendall"),
  to = c("pearson", "spearman", "kendall"))
}
\arguments{
\item{rho}{A a square symmetric correlation matrix}

\item{from}{The type of the input correlation matrix}

\item{to}{The type of the output correlation matrix}
}
\description{
Convert between different types of correlation matrices
}
