% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2.R
\name{convertKendall2Pearson}
\alias{convertKendall2Pearson}
\title{Converts Kendall correlation to Pearson. Used for normal random variables.}
\usage{
convertKendall2Pearson(rho)
}
\arguments{
\item{rho}{A a square symmetric Kendall correlation matrix.}
}
\value{
A Pearson correlation matrix.
}
\description{
Converts Kendall correlation to Pearson. Used for normal random variables.
}
