% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{convertKendallPearson}
\alias{convertKendallPearson}
\title{Converts Kendall correlation to Pearson. Used for normal random variables.}
\usage{
convertKendallPearson(rho)
}
\arguments{
\item{rho}{A a square symmetric Kendall correlation matrix.}
}
\value{
A Pearson correlation matrix.
}
\description{
Converts Kendall correlation to Pearson. Used for normal random variables.
}
