% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2.R
\name{convertPearson2Kendall}
\alias{convertPearson2Kendall}
\title{Converts Pearson correlation to Kendall. Used for normal random variables.}
\usage{
convertPearson2Kendall(rho)
}
\arguments{
\item{rho}{A a square symmetric Pearson correlation matrix.}
}
\value{
A Kendall correlation matrix.
}
\description{
Converts Pearson correlation to Kendall. Used for normal random variables.
}
