% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2.R
\name{convertPearson2Spearman}
\alias{convertPearson2Spearman}
\title{Converts Pearson correlation to Spearman. Used for normal random variables.}
\usage{
convertPearson2Spearman(rho)
}
\arguments{
\item{rho}{A a square symmetric Pearson correlation matrix.}
}
\value{
A Spearman correlation matrix.
}
\description{
Converts Pearson correlation to Spearman. Used for normal random variables.
}
