% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{convertPearsonKendall}
\alias{convertPearsonKendall}
\title{Converts Pearson correlation to Kendall. Used for normal random variables.}
\usage{
convertPearsonKendall(rho)
}
\arguments{
\item{rho}{A a square symmetric Pearson correlation matrix.}
}
\value{
A Kendall correlation matrix.
}
\description{
Converts Pearson correlation to Kendall. Used for normal random variables.
}
