% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{convertPearsonSpearman}
\alias{convertPearsonSpearman}
\title{Converts Pearson correlation to Spearman. Used for normal random variables.}
\usage{
convertPearsonSpearman(rho)
}
\arguments{
\item{rho}{A a square symmetric Pearson correlation matrix.}
}
\value{
A Spearman correlation matrix.
}
\description{
Converts Pearson correlation to Spearman. Used for normal random variables.
}
