% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2.R
\name{convertSpearman2Pearson}
\alias{convertSpearman2Pearson}
\title{Converts Spearman correlation to Pearson. Used for normal random variables.}
\usage{
convertSpearman2Pearson(rho)
}
\arguments{
\item{rho}{A a square symmetric Spearman correlation matrix.}
}
\value{
A Pearson correlation matrix.
}
\description{
Converts Spearman correlation to Pearson. Used for normal random variables.
}
