% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{convertSpearmanPearson}
\alias{convertSpearmanPearson}
\title{Converts Spearman correlation to Pearson. Used for normal random variables.}
\usage{
convertSpearmanPearson(Rho)
}
\arguments{
\item{R}{A a square symmetric Spearman correlation matrix.}
}
\value{
A Pearson correlation matrix.
}
\description{
Converts Spearman correlation to Pearson. Used for normal random variables.
}
