% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_bounds.R
\name{cor_bounds}
\alias{cor_bounds}
\title{Computes the theoretical upper and lower bounds of possible correlations
  given a set of marginals}
\usage{
cor_bounds(margins, method = c("pearson", "kendall", "spearman"), reps = 1e+05)
}
\arguments{
\item{margins}{A list containing the marginal distributions which are
distributions from the `distr6` package}

\item{method}{The type of correlation matrix that is being passed.}

\item{reps}{The number of sims used to estimate the bounds.}
}
\value{
A list containing the theoretical upper and lower bounds
}
\description{
Computes the theoretical upper and lower bounds of possible correlations
  given a set of marginals
}
