% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_convert.R
\name{cor_convert}
\alias{cor_convert}
\title{Convert from one correlation to another}
\usage{
cor_convert(
  rho,
  from = c("pearson", "spearman", "kendall"),
  to = c("pearson", "spearman", "kendall")
)
}
\arguments{
\item{rho}{A correlation matrix}

\item{from}{Input correlation matrix}

\item{to}{Output correlation matrix}
}
\description{
Convert from one correlation to another
}
