% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_fastPD.R
\name{cor_fastPD}
\alias{cor_fastPD}
\title{Return a positive definite correlation matrix that is close to the input matrix}
\usage{
cor_fastPD(R, tau = 1e-06)
}
\arguments{
\item{R}{The input correlation matrix}

\item{tau}{A minimum eigenvalue. Set to zero for positive semidefinite.}
}
\description{
Return a positive definite correlation matrix that is close to the input matrix
}
