% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_nearPD.R
\name{cor_nearPD}
\alias{cor_nearPD}
\title{Nearest positive (semi) definite correlation matrix}
\usage{
cor_nearPD(G, tau = 1e-06, tol = 1e-06)
}
\arguments{
\item{G}{the input correlation matrix}

\item{tau}{smallest allowed eigenvalue. Set to 0 for semidefinite matrix}

\item{tol}{relative gap between the objective function value of the
obtained solution and the objective function value of the global
solution.}
}
\description{
Nearest positive (semi) definite correlation matrix
}
