% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cor_nearPSD}
\alias{cor_nearPSD}
\title{Calculate the nearest positive semi-definite correlation matrix}
\usage{
cor_nearPSD(
  G,
  tau = 1e-05,
  iter_outer = 200L,
  iter_inner = 20L,
  maxit = 200L,
  err_tol = 1e-06,
  precg_err_tol = 0.01,
  newton_err_tol = 1e-04
)
}
\arguments{
\item{G}{the input correlation matrix}

\item{tau}{A user-dependent tuning parameter that determines the accuracy
of the final correlation matrix. Smaller values generally mean faster
convergence}

\item{iter_outer}{the max number of iterations in the outer loop}

\item{iter_inner}{the max number of iterations in the inner loop}

\item{maxit}{Maximum number of iterations in the pre_cg routine}

\item{err_tol}{the error tolerance for the stopping criteria}

\item{precg_err_tol}{the error tolerance in the pre-conjugate gradient method}

\item{newton_err_tol}{the error tolerance in Newton's method}
}
\description{
Calculate the nearest positive semi-definite correlation matrix
}
