% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_rand_.R
\name{cor_rand_PD}
\alias{cor_rand_PD}
\title{Compute a random positive definite correlation matrix}
\usage{
cor_rand_PD(d, a = 1)
}
\arguments{
\item{d}{A positive integer number of dimensions}

\item{a}{A tuning parameter}
}
\description{
Compute a random positive definite correlation matrix
}
