% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_rand_.R
\name{cor_rand_PSD}
\alias{cor_rand_PSD}
\title{Compute a random positive semi-definite correlation matrix}
\usage{
cor_rand_PSD(d, k = d)
}
\arguments{
\item{d}{A positive integer number of dimensions}

\item{k}{A tuning parameter}
}
\description{
Compute a random positive semi-definite correlation matrix
}
