% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_rand.R
\name{cor_randPD}
\alias{cor_randPD}
\title{Compute a random positive definite correlation matrix}
\usage{
cor_randPD(d, k = d)
}
\arguments{
\item{d}{A positive integer number of dimensions}

\item{k}{A tuning parameter between 1 and d}
}
\description{
Compute a random positive definite correlation matrix
}
