% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_rand.R
\name{cor_randPSD}
\alias{cor_randPSD}
\title{Compute a random positive semidefinite correlation matrix}
\usage{
cor_randPSD(d, k = d)
}
\arguments{
\item{d}{A positive integer number of dimensions}

\item{k}{A tuning parameter}
}
\description{
This method is generally faster than \code{\link{cor_randPD}}
}
