% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_rand.R
\name{.rjm}
\alias{.rjm}
\title{Internal function for `cor_rand_PD()`}
\usage{
.rjm(A, a)
}
\arguments{
\item{A}{The submatrix}

\item{a}{The parameters for `rbeta`}
}
\description{
Internal function for `cor_rand_PD()`
}
