% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_utils.R
\name{estimateSpearmanRho}
\alias{estimateSpearmanRho}
\title{Estimate the Spearman rank correlation}
\usage{
estimateSpearmanRho(x, y, fast = TRUE)
}
\description{
Estimate the Spearman rank correlation
}
