% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_utils.R
\name{is_valid_correlation}
\alias{is_valid_correlation}
\title{Check if a given correlation matrix is a valid}
\usage{
is_valid_correlation(rho)
}
\arguments{
\item{rho}{Input correlation matrix}
}
\description{
A correlation matrix is valid if:
* The diagonal elements are all equal to 1
* It is symmetric
* It is positive semi-definite
}
