% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jax_rand.R
\name{jax_rmvn}
\alias{jax_rmvn}
\title{Simulate random data from a multivariate normal distribution}
\usage{
jax_rmvn(n, mu, sigma)
}
\arguments{
\item{n}{The number of random vectors to be simulated.}

\item{mu}{A vector of length d, representing the mean.}

\item{sigma}{A symmetric positive definite covariance matrix (d x d)}
}
\description{
This function is designed to be 'bare bones' and does not check that the
  given covariance matrix is positive definite.
}
