% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mlist}
\alias{mlist}
\title{A helper function for creating a list of marginal distributions}
\usage{
mlist(...)
}
\arguments{
\item{...}{A list of marginal quantile functions with their named arguments}
}
\description{
Similar to `alist`, except that the right hand side of arguments are
    evaluated.
}
\examples{
x <- rnorm(500, 3.14, 1.2)
y <- rlnorm(500, 2.5, 1.0)
z <- rexp(500, 2.718)

m <- mlist(
  qnorm(mean = mean(x), sd = sd(x)),
  qlnorm(meanlog = mean(log(y)), sdlog = sd(log(y))),
  qexp(rate = (1 / mean(z)))
)

# will not work
# mlist(a = 10, b = 3)
}
