% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearPSDcor.R
\name{nearestPSDcor}
\alias{nearestPSDcor}
\title{Compute the nearest positive semidefinite correlation matrix}
\usage{
nearestPSDcor(
  R,
  tau = 1e-05,
  iter_outer = 200L,
  iter_inner = 20L,
  N = 200L,
  err_tol = 1e-06,
  precg_err_tol = 0.01,
  newton_err_tol = 1e-04
)
}
\arguments{
\item{R}{the input correlation matrix}

\item{tau}{some non-negative threshold for the smallest eigenvalue}

\item{iter_outer}{the max number of iterations in the outer loop}

\item{iter_inner}{the max number of iterations in the inner loop}

\item{N}{the max number of iterations in the pre-conjugate gradient method}

\item{err_tol}{the error tolerance for the stopping criteria}

\item{precg_err_tol}{the error tolerance in the pre-conjugate gradient method}

\item{newton_err_tol}{the error tolerance in Newton's method}
}
\description{
Compute the nearest positive semidefinite correlation matrix
}
