% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{normal2marginal}
\alias{normal2marginal}
\title{Transforms a [multivariate]normal vector to a different marginal via a
uniform intermediate transformation.}
\usage{
normal2marginal(x, param)
}
\arguments{
\item{x}{A normal random vector.}

\item{param}{A list ccontaining the marginal and its parameters.}
}
\description{
Transforms a [multivariate]normal vector to a different marginal via a
uniform intermediate transformation.
}
