% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numpy_hermite_physicists.R
\name{numpy_hermgauss}
\alias{numpy_hermgauss}
\title{Gauss-Hermite quadrature - "Physicists"}
\usage{
numpy_hermgauss(n = 1L)
}
\arguments{
\item{n}{Number of sample points and weights. It must be >= 1.}
}
\value{
A list of nodes and weights
}
\description{
Computes the sample points and weights for Gauss-Hermite quadrature. These
    sample points and weights will correctly integrate polynomials of degree
    \eqn{2 * n - 1} or less over the interval \eqn{[-Inf, Inf]} with the weight
    function \eqn{f(x) = \exp(-x^2)}.
}
