% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numpy_hermite_probabilists.R
\name{numpy_hermite_e}
\alias{numpy_hermite_e}
\title{Hermite polynomial evaluation - "Probabilists"}
\usage{
numpy_hermite_e(x, n)
}
\arguments{
\item{x}{a scalar or vector of numerical values}

\item{n}{The degree of the Hermite polynomial}
}
\description{
Evaluate the nth degree Hermite polynomial at x
}
