% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_bounds.R
\name{pearson_bounds}
\alias{pearson_bounds}
\title{Compute the theoretical lower and upper Pearson correlation bounds}
\usage{
pearson_bounds(dA, dB, n = 7)
}
\arguments{
\item{dA}{The first marginal distribution from the `distr6` package}

\item{dB}{The second marginal distribution from the `distr6` package}

\item{n}{The degree of the Hermite polynomial}
}
\description{
Compute the theoretical lower and upper Pearson correlation bounds
}
