% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_matching.R
\name{pm_discrete_case}
\alias{pm_discrete_case}
\title{Pearson matching for the discrete-discrete case}
\usage{
pm_discrete_case(r, dA, dB, n)
}
\arguments{
\item{r}{The target correlation}

\item{dA}{The first marginal distribution}

\item{dB}{The second marginal distribution}

\item{n}{The degree of the Hermite polynomial}
}
\description{
Pearson matching for the discrete-discrete case
}
