% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crude_helper_funcs.R
\name{rcorr}
\alias{rcorr}
\title{Generate a random Pearson correlation matrix}
\usage{
rcorr(d, constant_rho = FALSE)
}
\arguments{
\item{d}{A positive integer.}

\item{constant_rho}{A boolean value.}
}
\value{
A \code{dxd} symmetric matrix with ones on the diagonal.
}
\description{
Generate a random Pearson correlation matrix
}
\examples{
rcorr(4)
rcorr(10, TRUE)
}
