% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvn.R
\name{rmvn}
\alias{rmvn}
\title{Simulate random data from a multivariate normal distribution}
\usage{
rmvn(n, mu, sigma, isChol = FALSE)
}
\arguments{
\item{n}{number of random vectors to be simulated.}

\item{mu}{vector of length d, representing the mean.}

\item{sigma}{covariance matrix (d x d). Alternatively is can be the cholesky
decomposition of the covariance. In that case isChol should be set to TRUE.}

\item{isChol}{boolean set to true if sigma is the cholesky decomposition of
the covariance matrix.}
}
\description{
This function is designed to be 'bare bones' and does not check that the
  given covariance matrix is positive semi-definite.
}
