% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvu.R
\name{rmvu}
\alias{rmvu}
\title{Generate random data from a multivariate uniform distribution}
\usage{
rmvu(n, rho, min = 0, max = 1)
}
\arguments{
\item{n}{number of random vectors to be simulated}

\item{rho}{correlation matrix (d x d)}

\item{min}{either a single number or a vector of length d}

\item{max}{either a single number or a vector of length d}
}
\description{
Generate random data from a multivariate uniform distribution
}
