% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crude_helper_funcs.R
\name{rnegbin_params}
\alias{rnegbin_params}
\title{Generate negative binomial marginal parameters}
\usage{
rnegbin_params(d, shape = 100, id_margins = FALSE)
}
\arguments{
\item{d}{A positive integer.}

\item{shape}{The shape parameter of a negative binomial distribution.}

\item{id_margins}{A boolean value specifying if the margins should be identical.}
}
\value{
A matrix with columns containing (in order) lambda, alpha, mean, and variance.
}
\description{
Generate negative binomial marginal parameters
}
\examples{
rnegbin_params(4)
rnegbin_params(10, shape = 40)
rnegbin_params(7, 44, TRUE)
}
