% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvec.R
\name{rvec}
\alias{rvec}
\title{Creates \code{n} observations from a multivariate distribution with the
given marginals and correlation matrix.}
\usage{
rvec(n, R, params, cores = 1, type = c("pearson", "kendall",
  "spearman"))
}
\arguments{
\item{n}{The number random vectors to generate.}

\item{R}{The input correlation matrix.}

\item{params}{The parameters of the marginals.}

\item{cores}{The number of cores to utilize.}

\item{type}{The type of correlation matrix that is being passed.}
}
\value{
A matrix of random vectors generated from the specified marginals and parameters.
}
\description{
Creates \code{n} observations from a multivariate distribution with the
given marginals and correlation matrix.
}
\examples{
d <- 5
rho <- rcor(d)
margins <- list(
  list("nbinom", size = 5, prob = 0.3),
  list("exp", rate = 4),
  list("binom", size = 5, prob = 0.7),
  list("norm", mean = 10, sd = 3),
  list("pois", lambda = 10)
)

margins2 <- list(
  list("nbinom", 5, 0.3),
  list("exp", 4),
  list("binom", 5, 0.7),
  list("norm", 10, 3),
  list("pois", 10)
)

rvec(10, rho, margins, cores = 1, type = "pearson")
rvec(10, rho, margins2, cores = 1, type = "pearson")
}
