% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simInvNB.R
\name{simInvNB}
\alias{simInvNB}
\title{Creates \code{n} observations from a multivariate negative binomial distribution.}
\usage{
simInvNB(n, R, params, cores = 1, type = c("pearson", "kendall",
  "spearman"))
}
\arguments{
\item{n}{The number random vectors to generate.}

\item{R}{The input correlation matrix.}

\item{params}{The parameters of the marginals.}

\item{cores}{The number of cores to utilize.}

\item{type}{The type of correlation matrix that is being passed.}
}
\description{
Creates \code{n} observations from a multivariate negative binomial distribution.
}
