% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_utils.R
\name{which_corInBounds}
\alias{which_corInBounds}
\title{Returns a logical matrix of which correlations are in the feasible region}
\usage{
which_corInBounds(rho, rho_bounds, negate = FALSE)
}
\arguments{
\item{rho}{The input correlation matrix.}

\item{rho_bounds}{A list containing the theoretical upper and lower bounds}

\item{negate}{Should the logical values be negated in order to identify
values that are outside the feasible region.}
}
\description{
Returns a logical matrix of which correlations are in the feasible region
}
