% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCorBounds.R
\name{computeCorBounds}
\alias{computeCorBounds}
\title{Computes the theoretical upper and lower bounds of possible correlations
  given a set of marginals}
\usage{
computeCorBounds(
  margins,
  type = c("pearson", "kendall", "spearman"),
  cores = 1,
  reps = 1e+05
)
}
\arguments{
\item{margins}{The parameters of the marginals.}

\item{type}{The type of correlation matrix that is being passed.}

\item{cores}{The number of cores to utilize.}

\item{reps}{The number of reps used for computing the cor bounds}
}
\value{
A list containing the theoretical upper and lower bounds
}
\description{
Computes the theoretical upper and lower bounds of possible correlations
  given a set of marginals
}
