% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{distributions_setup}
\alias{distributions_setup}
\title{Setup Distributions.jl}
\usage{
distributions_setup()
}
\value{
Return the imported wrapper of Distributions.jl Julia package
}
\description{
This function initializes the Distributions package that many of the Bigsimr
functions work with.
}
