% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_utils.R
\name{estimateSpearmanRho}
\alias{estimateSpearmanRho}
\title{Estimate the Spearman rank correlation}
\usage{
estimateSpearmanRho(x, fast = TRUE)
}
\description{
For a Nx2 matrix when N=1e5, fast is about 2x faster, and when N=1e6,
  fast is about 3x faster.
}
