% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_utils.R
\name{rnbinom_params}
\alias{rnbinom_params}
\title{Generate negative binomial marginal parameters}
\usage{
rnbinom_params(d, id_margins = FALSE)
}
\arguments{
\item{d}{A positive integer.}

\item{id_margins}{A boolean value specifying if the margins should be
identical.}
}
\value{
A matrix with columns containing (in order) lambda, alpha, mean,
  and variance.
}
\description{
Generate negative binomial marginal parameters
}
\examples{
rnbinom_params(4)
rnbinom_params(10)
rnbinom_params(7, TRUE)
}
