% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_vec.R
\name{rvec}
\alias{rvec}
\title{Creates \code{n} observations from a multivariate distribution with the
given marginals and correlation matrix.}
\usage{
rvec(
  n,
  rho,
  margins,
  type = c("pearson", "kendall", "spearman"),
  ensure_PSD = FALSE,
  cores = 1L
)
}
\arguments{
\item{n}{The number random vectors to generate.}

\item{rho}{The input correlation matrix.}

\item{margins}{The marginal distributions (Typically R's "quantile functions")}

\item{type}{The type of correlation matrix that is being passed (Assumed to
be Pearson by default).}

\item{ensure_PSD}{Ensure that the converted correlation matrix is positive
semi-definite. More import if the input correlation type is Kendall or
Spearman.}

\item{cores}{The number of cores to run on}
}
\value{
A matrix of random vectors generated from the specified marginals
  and parameters.
}
\description{
Creates \code{n} observations from a multivariate distribution with the
given marginals and correlation matrix.
}
