% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvec.R
\name{rvec}
\alias{rvec}
\title{Creates \code{n} observations from a multivariate distribution with the
given marginals and correlation matrix.}
\usage{
rvec(n, rho, margins, cores = 1, type = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{n}{The number random vectors to generate.}

\item{rho}{The input correlation matrix.}

\item{margins}{The marginal distributions (Typically R's "quantile functions")}

\item{cores}{The number of cores to utilize.}

\item{type}{The type of correlation matrix that is being passed (Assumed to
be Pearson by default).}
}
\value{
A matrix of random vectors generated from the specified marginals
  and parameters.
}
\description{
Creates \code{n} observations from a multivariate distribution with the
given marginals and correlation matrix.
}
