% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_vec.R
\name{rvec}
\alias{rvec}
\title{Simulate correlated multivariate data}
\usage{
rvec(n, rho, margins, cores = 1L, ensure_PSD = FALSE)
}
\arguments{
\item{n}{The number random vectors to generate.}

\item{rho}{The input (Pearson) correlation matrix.}

\item{margins}{A list containing the marginal distributions which are
distributions from the `distr6` package}

\item{cores}{The number of cores to run on.}

\item{ensure_PSD}{Ensure that the converted correlation matrix is positive
semi-definite.}
}
\value{
A matrix of random vectors generated from the specified marginals
  and parameters.
}
\description{
Creates \code{n} observations from a multivariate distribution with the
  given marginals and correlation matrix. The correlation matrix is always
  assumed to be a Pearson correlation matrix
}
